Merging data frames and storing output suggestions
What are the sizes of the BrapaP and BrapaP_male dataframes when you are trying to merge them? Are they 11,000 and 1,500 respectively? Or are you trying to merge only a portion of each dataframe? I noticed that you have all.x=T. Do you want this? Because that will keep all rows regardless of if they have a match in the 'id' column or not. If your extra columns have a bunch of NAs then that's a sign that this is what is wrong. If you want to retain only rows that have a corresponding row in BOTH data frames, all.x should = F. Sorry if this is not helpful!